Changed maxWidth and maxHeight on editTimeReport and createTimeReport in epuss.js. Changed type from int to long in Number. Did some refactoring in WorkerComponent and removed comments in SerletBase.

+++ src/data/Number.java
@@ -10,13 +10,13 @@ public enum Number {
     SDP(11), SRS(12), SVVS(13), STLDD(14), SVVI(15), SDDD(16), SVVR(17), SSD(18), 
     ProjectFinalReport(19), FunctionalTesting(21), SystemTesting(22), RegressionTesting(23), Meeting(30),
     Lecture(41), Exercise(42), ComputerExercise(43), OutsideReading(44), Other(100);
-    private int value;
+    private long value;
 
     private Number(int value) {
     	this.value = value;
     }
     
-    public static boolean isNumber(int number){
+    public static boolean isNumber(long number){
     	boolean isType = false;
     	for (Number t : Number.values()){
     		if(t.value == number){
@@ -26,7 +26,7 @@ public enum Number {
     	return isType;
     }
     
-    public int getValue() {
+    public long getValue() {
     	return this.value;
     }
 };
